use hypotf

+++ src/minimodem.c
@@ -31,7 +31,7 @@ band_mag( fftwf_complex * const cplx, unsigned int band, float scalar )
 {
     float re = cplx[band][0];
     float im = cplx[band][1];
-    float mag = hypot(re, im);
+    float mag = hypotf(re, im);
     return mag * scalar;
 }
